requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
librosa>=0.10.0
soundfile>=0.12.0
jiwer>=3.0.0
gradio>=4.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
